changed price to decimal

+++ db/migrate/20160112215315_add_price_to_orders.rb
@@ -1,5 +1,5 @@
 class AddPriceToOrders < ActiveRecord::Migration
   def change
-    add_column :orders, :price, :integer
+    add_column :orders, :price, :decimal
   end
 end
